Move stats utilities to utils and update imports
